requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
